usp-backend/pyproject.toml
[tool.poetry]
name = "usp-backend"
version = "0.1.0"
description = ""
authors = ["tkk <t.kleylein@startmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlmodel = "0.0.8"
sqlalchemy = "^1.4.41"
semver = "^3.0.1"
requests = "^2.31.0"
python-multipart = "^0.0.6"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
toml = "^0.10.2"
psycopg2-binary = "^2.9.7"
alembic = "^1.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
